+++ src/nativeInterop/cinterop/intrin/ssse3.cpp
@@ -50,7 +50,7 @@ namespace alignr_epi8 {
 		{32, [](__m128i a, __m128i b) { return _mm_alignr_epi8(a, b, 32); }},
 	};
 } // namespace alignr_epi8
-__m128i ssse3_alignr_epi8(__m128i a, __m128i b, int8_t imm8) { if (imm8 > 32) imm8 = 32; return alignr_epi8::map.at(imm8)(a, b); }
+__m128i ssse3_alignr_epi8(__m128i a, __m128i b, int32_t imm8) { if (imm8 > 32) imm8 = 32; return alignr_epi8::map.at(imm8)(a, b); }
 __m128i ssse3_hadd_epi16(__m128i a, __m128i b) { return _mm_hadd_epi16(a, b); }
 __m128i ssse3_hadd_epi32(__m128i a, __m128i b) { return _mm_hadd_epi32(a, b); }
 __m128i ssse3_hadds_epi16(__m128i a, __m128i b) { return _mm_hadds_epi16(a, b); }
